refactor(governance): split package into root types + service subpackage
Follow the standard service-package layout: domain types in governance.go,
the Service interface in service.go, and the implementation (Config, New,
QueryAccess orchestration, entitlement→access mapping) under service/.

Tests move alongside the implementation in service/. Wiring updated to call
governanceservice.New; provider signature is unchanged so no wire regen.
