+++ openmeter/governance/service/service_test.go
@@ -1,4 +1,4 @@
-package governance
+package service
 
 import (
 	"context"
@@ -18,6 +18,7 @@ import (
 	customerservice "github.com/openmeterio/openmeter/openmeter/customer/service"
 	entdb "github.com/openmeterio/openmeter/openmeter/ent/db"
 	"github.com/openmeterio/openmeter/openmeter/entitlement"
+	"github.com/openmeterio/openmeter/openmeter/governance"
 	"github.com/openmeterio/openmeter/openmeter/meter"
 	meteradapter "github.com/openmeterio/openmeter/openmeter/meter/mockadapter"
 	"github.com/openmeterio/openmeter/openmeter/productcatalog/feature"
@@ -128,7 +129,7 @@ func setupTestDeps(t *testing.T) *testDeps {
 	}
 }
 
-func newTestService(t *testing.T, deps *testDeps) Service {
+func newTestService(t *testing.T, deps *testDeps) governance.Service {
 	t.Helper()
 	svc, err := New(Config{
 		CustomerService:    deps.customerService,
@@ -259,7 +260,7 @@ func TestQueryAccess_UnknownCustomerKey(t *testing.T) {
 	svc := newTestService(t, deps)
 	ns := newTestNamespace(t)
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"ghost"},
 		PageSize:     testPageSize,
@@ -267,7 +268,7 @@ func TestQueryAccess_UnknownCustomerKey(t *testing.T) {
 	require.NoError(t, err)
 	assert.Empty(t, res.Customers)
 	require.Len(t, res.Errors, 1)
-	assert.Equal(t, QueryErrorCustomerNotFound, res.Errors[0].Code)
+	assert.Equal(t, governance.QueryErrorCustomerNotFound, res.Errors[0].Code)
 	assert.Equal(t, "ghost", res.Errors[0].CustomerKey)
 }
 
@@ -280,7 +281,7 @@ func TestQueryAccess_KnownCustomerNoEntitlements(t *testing.T) {
 
 	cust := createCustomer(t, deps, ns, "acme", []string{"acme"})
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{cust.GetUsageAttribution().SubjectKeys[0]},
 		PageSize:     testPageSize,
@@ -301,7 +302,7 @@ func TestQueryAccess_BooleanEntitlement_HasAccess(t *testing.T) {
 	cust := createCustomer(t, deps, ns, "acme", []string{"acme"})
 	createBooleanFeatureAndEntitlement(t, deps, ns, "premium", cust)
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme"},
 		FeatureKeys:  []string{"premium"},
@@ -325,7 +326,7 @@ func TestQueryAccess_FeatureNotFound(t *testing.T) {
 
 	createCustomer(t, deps, ns, "acme", []string{"acme"})
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme"},
 		FeatureKeys:  []string{"does-not-exist"},
@@ -336,7 +337,7 @@ func TestQueryAccess_FeatureNotFound(t *testing.T) {
 	fa := res.Customers[0].Features["does-not-exist"]
 	assert.False(t, fa.HasAccess)
 	require.NotNil(t, fa.Reason)
-	assert.Equal(t, ReasonFeatureNotFound, fa.Reason.Code)
+	assert.Equal(t, governance.ReasonFeatureNotFound, fa.Reason.Code)
 }
 
 func TestQueryAccess_FeatureUnavailable(t *testing.T) {
@@ -350,7 +351,7 @@ func TestQueryAccess_FeatureUnavailable(t *testing.T) {
 	createCustomer(t, deps, ns, "acme", []string{"acme"})
 	createOrphanFeature(t, deps, ns, "enterprise")
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme"},
 		FeatureKeys:  []string{"enterprise"},
@@ -361,7 +362,7 @@ func TestQueryAccess_FeatureUnavailable(t *testing.T) {
 	fa := res.Customers[0].Features["enterprise"]
 	assert.False(t, fa.HasAccess)
 	require.NotNil(t, fa.Reason)
-	assert.Equal(t, ReasonFeatureUnavailable, fa.Reason.Code)
+	assert.Equal(t, governance.ReasonFeatureUnavailable, fa.Reason.Code)
 }
 
 func TestQueryAccess_MultipleKeysSameCustomer(t *testing.T) {
@@ -375,7 +376,7 @@ func TestQueryAccess_MultipleKeysSameCustomer(t *testing.T) {
 	// customer key = "acme", usage attribution subject key = "acme-sub"
 	createCustomer(t, deps, ns, "acme", []string{"acme-sub"})
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme", "acme-sub"},
 		PageSize:     testPageSize,
@@ -397,7 +398,7 @@ func TestQueryAccess_MixedHitsAndMisses(t *testing.T) {
 	cust := createCustomer(t, deps, ns, "acme", []string{"acme"})
 	createBooleanFeatureAndEntitlement(t, deps, ns, "feature-a", cust)
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme", "unknown-key"},
 		FeatureKeys:  []string{"feature-a"},
@@ -406,7 +407,7 @@ func TestQueryAccess_MixedHitsAndMisses(t *testing.T) {
 	require.NoError(t, err)
 	require.Len(t, res.Customers, 1)
 	require.Len(t, res.Errors, 1)
-	assert.Equal(t, QueryErrorCustomerNotFound, res.Errors[0].Code)
+	assert.Equal(t, governance.QueryErrorCustomerNotFound, res.Errors[0].Code)
 	assert.True(t, res.Customers[0].Features["feature-a"].HasAccess)
 }
 
@@ -425,7 +426,7 @@ func TestQueryAccess_NoFeatureKeysReturnsAll(t *testing.T) {
 	// feat-3 exists in the org but the customer has no entitlement for it.
 	createOrphanFeature(t, deps, ns, "feat-3")
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme"},
 		PageSize:     testPageSize,
@@ -438,7 +439,7 @@ func TestQueryAccess_NoFeatureKeysReturnsAll(t *testing.T) {
 	feat3 := res.Customers[0].Features["feat-3"]
 	assert.False(t, feat3.HasAccess)
 	require.NotNil(t, feat3.Reason)
-	assert.Equal(t, ReasonFeatureUnavailable, feat3.Reason.Code)
+	assert.Equal(t, governance.ReasonFeatureUnavailable, feat3.Reason.Code)
 }
 
 func TestQueryAccess_MeteredEntitlement_HasAccess(t *testing.T) {
@@ -462,7 +463,7 @@ func TestQueryAccess_MeteredEntitlement_HasAccess(t *testing.T) {
 
 	clock.SetTime(now.Add(time.Hour))
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme"},
 		FeatureKeys:  []string{"premium"},
@@ -496,7 +497,7 @@ func TestQueryAccess_MeteredEntitlement_Exhausted(t *testing.T) {
 
 	clock.SetTime(now.Add(time.Hour))
 
-	res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 		Namespace:    ns,
 		CustomerKeys: []string{"acme"},
 		FeatureKeys:  []string{"premium"},
@@ -508,7 +509,7 @@ func TestQueryAccess_MeteredEntitlement_Exhausted(t *testing.T) {
 	fa := res.Customers[0].Features["premium"]
 	assert.False(t, fa.HasAccess)
 	require.NotNil(t, fa.Reason)
-	assert.Equal(t, ReasonUsageLimitReached, fa.Reason.Code)
+	assert.Equal(t, governance.ReasonUsageLimitReached, fa.Reason.Code)
 }
 
 func TestQueryAccess_Pagination(t *testing.T) {
@@ -531,8 +532,8 @@ func TestQueryAccess_Pagination(t *testing.T) {
 
 	allKeys := []string{"c1", "c2", "c3"}
 
-	query := func(after, before *pagination.Cursor) QueryResult {
-		res, err := svc.QueryAccess(t.Context(), QueryAccessInput{
+	query := func(after, before *pagination.Cursor) governance.QueryResult {
+		res, err := svc.QueryAccess(t.Context(), governance.QueryAccessInput{
 			Namespace:    ns,
 			CustomerKeys: allKeys,
 			PageSize:     1,
